third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-control-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
CONSOLE MESSAGE: PASS: Iframe was not in a unique origin
